+++ lua/sniprun/display.lua
@@ -5,9 +5,9 @@ M.term.buffer = -1
 M.term.window_handle = 0
 M.term.current_line = -1
 M.term.chan = -1
-M.borders = 'single'
+M.borders = "single"
 
-local NAMESPACE = 'sniprun'
+local NAMESPACE = "sniprun"
 
 function M.fw_open(row, column, message, ok, temp)
     M.fw_close()
@@ -29,23 +29,23 @@ function M.fw_open(row, column, message, ok, temp)
         vim.api.nvim_buf_add_highlight(bufnr, namespace_id, hl, h, 0, -1) -- highlight lines in floating window
     end
     if h ~= 0 then
-        M.fw_handle = vim.api.nvim_open_win(bufnr, false,
-            {
-                relative = 'win',
-                width = w + 1,
-                height = h,
-                bufpos = bp,
-                focusable = false,
-                style = 'minimal',
-                border = M
-                    .borders
-            })
+        M.fw_handle = vim.api.nvim_open_win(bufnr, false, {
+            relative = "win",
+            width = w + 1,
+            height = h,
+            bufpos = bp,
+            focusable = false,
+            style = "minimal",
+            border = M.borders,
+        })
     end
 end
 
 function M.term_set_window_handle()
     local winid = vim.fn.bufwinid(M.term.buffer)
-    if winid ~= -1 then return end
+    if winid ~= -1 then
+        return
+    end
 
     local location = require("sniprun").config_values.display_options.terminal_position
     if location == "horizontal" then
@@ -109,13 +109,12 @@ function M.write_to_term(message, ok)
     -- the window screen minus the length of the message and the two spaces on
     -- each side.
     local numdashes = (wininfo.width - wininfo.textoff - status:len() - 4) / 2
-
     -- If the status message is not even, then 'numdashes' will be a float, and
     -- the header has wrong number of dashes. So we round it down as the prefix,
     -- and up for the suffix.
 
-    header_prefix = string.rep("-", math.floor(numdashes))
-    header_suffix = string.rep("-", math.ceil(numdashes))
+    local header_prefix = string.rep("-", math.floor(numdashes))
+    local header_suffix = string.rep("-", math.ceil(numdashes))
 
     -- It's valid for the message to contain null characters per the neovim
     -- specification, so we try to avoid performing string operations on it by
@@ -134,18 +133,23 @@ function M.close_all()
 end
 
 function M.fw_close()
-    if M.fw_handle == 0 then return end
+    if M.fw_handle == 0 then
+        return
+    end
     vim.api.nvim_win_close(M.fw_handle, true)
     M.fw_handle = 0
 end
 
 function M.clear_virtual_text()
     vim.cmd(
-        "let sniprun_namespace_id = nvim_create_namespace('sniprun')\n call nvim_buf_clear_namespace(0,sniprun_namespace_id, 0 ,-1)")
+        "let sniprun_namespace_id = nvim_create_namespace('sniprun')\n call nvim_buf_clear_namespace(0,sniprun_namespace_id, 0 ,-1)"
+    )
 end
 
 function M.term_close()
-    if M.term.window_handle == 0 then return end
+    if M.term.window_handle == 0 then
+        return
+    end
     vim.api.nvim_win_close(M.term.window_handle, true)
     M.term.window_handle = 0
     M.term.buffer = -1
@@ -157,26 +161,39 @@ function M.display_nvim_notify(message, ok)
     -- ok is a boolean variable for the status (true= ok, false= error)
     --
     -- test if nvim_notify is availablea
-    if pcall(function () require('notify') end) then
+    if pcall(function()
+        require("notify")
+    end) then
         --ok
     else
         print("Sniprun: nvim_notify is not installed")
         return
     end
 
-    if message == "" then return end
+    if message == "" then
+        return
+    end
     local title = ok and "Sniprun: Ok" or "Sniprun: Error"
     local notif_style = ok and "info" or "error"
-    require("notify")(message, notif_style,
-        {
-            title = title,
-            timeout = require('sniprun').config_values.display_options.notification_timeout * 1000,
-            render = require('sniprun').config_values.display_options.notification_render or "default"
-        })
+    require("notify")(message, notif_style, {
+        title = title,
+        timeout = require("sniprun").config_values.display_options.notification_timeout * 1000,
+        render = require("sniprun").config_values.display_options.notification_render or "default",
+    })
 end
 
-function M.display_extmark(ns, line, message, highlight)
+function M.display_virt_text(ns, line, message, highlight)
     vim.api.nvim_buf_set_extmark(0, ns, line, -1, { virt_text = { { message, highlight } } })
+end
+
+function M.display_virt_line(ns, line_pos, message, highlight)
+    local virt_lines = {}
+    for line in message:gmatch("([^\n]*)\n?") do
+        table.insert(virt_lines, { { line, highlight } })
+    end
+    vim.api.nvim_buf_set_extmark(0, ns, line_pos, 0, {
+        virt_lines = virt_lines,
+    })
 end
 
 function M.send_api(message, ok)
@@ -187,7 +204,7 @@ function M.send_api(message, ok)
     else
         d.status = "error"
     end
-    local listeners = require('sniprun.api').listeners
+    local listeners = require("sniprun.api").listeners
 
     if type(next(listeners)) == "nil" then
         print("Sniprun: No listener registered")
@@ -199,7 +216,7 @@ function M.send_api(message, ok)
 end
 
 function M.close_api()
-    local listeners = require('sniprun.api').closers
+    local listeners = require("sniprun.api").closers
     for i, f in ipairs(listeners) do
         f()
     end