+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v1.3.17
 - better fallback messages
+- Terminal display mode fixes (courtesy of @dbeecham)
 
 ## v1.3.16
 - Fix nvim-notify timeout and add render style option