+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.3.17
+- better fallback messages
+- Terminal display mode fixes (courtesy of @dbeecham)
+- VirtualLine display mode
+
 ## v1.3.16
 - Fix nvim-notify timeout and add render style option
 - Filter ANSI escape code by default